Add support for the system backend
Previously vperfetto assumed the inprocess backend.

Now it can use the system backend.

This is done through changing the argument to startTracing to a config struct.

The config struct contains fields for the filename and flags for whether to
use the in-process or system backend.

The filename is not needed when using the system backend.

Also add a validation function to check the config is good before enabling tracing.